+++ experiment_manager/job/job.py
@@ -219,9 +219,10 @@ class Job(object):
 		self.clean_backup()
 		os.makedirs(own_backup_dir)
 		for f in self.files:
-			if not os.path.isdir(os.path.dirname(os.path.join(own_backup_dir,f))):
-				os.makedirs(os.path.dirname(os.path.join(own_backup_dir,f)))
-			shutil.copy(f,os.path.join(own_backup_dir,f))
+			if os.path.isfile(f)
+				if not os.path.isdir(os.path.dirname(os.path.join(own_backup_dir,f))):
+					os.makedirs(os.path.dirname(os.path.join(own_backup_dir,f)))
+				shutil.copy(f,os.path.join(own_backup_dir,f))
 		#shutil.copytree('.',own_backup_dir,ignore=shutil.ignore_patterns(own_backup_dir))
 		os.remove(backup_lock_file)
